+++ pkg/openfeature/event_executor.go
@@ -1,56 +1,51 @@
 package openfeature
 
 import (
+	"context"
 	"fmt"
 	"sync"
 	"time"
 
 	"github.com/go-logr/logr"
+	"golang.org/x/sync/errgroup"
 )
 
 // event executor is a registry to connect API and Client event handlers to Providers
 
+// handlerExecutionTime defines the maximum time event handler will wait for its handlers to complete
+const handlerExecutionTime = 500 * time.Millisecond
+
 type eventExecutor struct {
-	defaultProviderReference *providerReference
-	namedProviderReference   map[string]*providerReference
-	apiRegistry              map[EventType][]EventCallback
-	scopedRegistry           map[string]scopedCallback
-	logger                   logr.Logger
-	mu                       sync.Mutex
+	providerShutdownHook map[string]chan interface{}
+	apiRegistry          map[EventType][]EventCallBack
+	scopedRegistry       map[string]scopedCallback
+	logger               logr.Logger
+	mu                   sync.Mutex
 }
 
-func newEventExecutor(logger logr.Logger) eventExecutor {
+func newEventHandler(logger logr.Logger) eventExecutor {
 	return eventExecutor{
-		namedProviderReference: map[string]*providerReference{},
-		apiRegistry:            map[EventType][]EventCallback{},
-		scopedRegistry:         map[string]scopedCallback{},
-		logger:                 logger,
+		providerShutdownHook: map[string]chan interface{}{},
+		apiRegistry:          map[EventType][]EventCallBack{},
+		scopedRegistry:       map[string]scopedCallback{},
+		logger:               logger,
 	}
 }
 
-// scopedCallback is a helper struct to hold client name associated callbacks.
+// scopedCallback is a helper struct to hold scope specific callbacks.
 // Here, the scope correlates to the client and provider name
 type scopedCallback struct {
 	scope     string
-	callbacks map[EventType][]EventCallback
+	callbacks map[EventType][]EventCallBack
 }
 
 func newScopedCallback(client string) scopedCallback {
 	return scopedCallback{
 		scope:     client,
-		callbacks: map[EventType][]EventCallback{},
+		callbacks: map[EventType][]EventCallBack{},
 	}
 }
 
-type providerReference struct {
-	eventHandler          *EventHandler
-	metadata              Metadata
-	clientNameAssociation string
-	isDefault             bool
-	shutdownSemaphore     chan interface{}
-}
-
-// updateLogger updates the executor's logger
 func (e *eventExecutor) updateLogger(l logr.Logger) {
 	e.mu.Lock()
 	defer e.mu.Unlock()
@@ -58,20 +53,18 @@ func (e *eventExecutor) updateLogger(l logr.Logger) {
 	e.logger = l
 }
 
-// registerApiHandler adds an API(global) level handler
-func (e *eventExecutor) registerApiHandler(t EventType, c EventCallback) {
+func (e *eventExecutor) registerApiHandler(t EventType, c EventCallBack) {
 	e.mu.Lock()
 	defer e.mu.Unlock()
 
 	if e.apiRegistry[t] == nil {
-		e.apiRegistry[t] = []EventCallback{c}
+		e.apiRegistry[t] = []EventCallBack{c}
 	} else {
 		e.apiRegistry[t] = append(e.apiRegistry[t], c)
 	}
 }
 
-// removeApiHandler removes an API(global) level handler
-func (e *eventExecutor) removeApiHandler(t EventType, c EventCallback) {
+func (e *eventExecutor) removeApiHandler(t EventType, c EventCallBack) {
 	e.mu.Lock()
 	defer e.mu.Unlock()
 
@@ -90,8 +83,7 @@ func (e *eventExecutor) removeApiHandler(t EventType, c EventCallback) {
 	e.apiRegistry[t] = entrySlice
 }
 
-// registerClientHandler registers a client level handler
-func (e *eventExecutor) registerClientHandler(clientName string, t EventType, c EventCallback) {
+func (e *eventExecutor) registerClientHandler(clientName string, t EventType, c EventCallBack) {
 	e.mu.Lock()
 	defer e.mu.Unlock()
 
@@ -100,41 +92,16 @@ func (e *eventExecutor) registerClientHandler(clientName string, t EventType, c 
 		e.scopedRegistry[clientName] = newScopedCallback(clientName)
 	}
 
-	registry := e.scopedRegistry[clientName]
+	callback := e.scopedRegistry[clientName]
 
-	if registry.callbacks[t] == nil {
-		registry.callbacks[t] = []EventCallback{c}
+	if callback.callbacks[t] == nil {
+		callback.callbacks[t] = []EventCallBack{c}
 	} else {
-		registry.callbacks[t] = append(registry.callbacks[t], c)
-	}
-
-	// fulfil spec requirement to fire ready events if associated provider is ready
-	if t != ProviderReady {
-		return
-	}
-
-	provider, ok := e.namedProviderReference[clientName]
-	if !ok {
-		return
-	}
-
-	s, ok := (*provider.eventHandler).(StateHandler)
-	if !ok {
-		return
-	}
-
-	if s.Status() == ReadyState {
-		(*c)(EventDetails{
-			providerName: provider.metadata.Name,
-			ProviderEventDetails: ProviderEventDetails{
-				Message: "provider is at ready state",
-			},
-		})
+		callback.callbacks[t] = append(callback.callbacks[t], c)
 	}
 }
 
-// removeClientHandler removes a client level handler
-func (e *eventExecutor) removeClientHandler(name string, t EventType, c EventCallback) {
+func (e *eventExecutor) removeClientHandler(name string, t EventType, c EventCallBack) {
 	e.mu.Lock()
 	defer e.mu.Unlock()
 
@@ -159,138 +126,106 @@ func (e *eventExecutor) removeClientHandler(name string, t EventType, c EventCal
 	e.scopedRegistry[name].callbacks[t] = entrySlice
 }
 
-// registerDefaultProvider registers the default FeatureProvider and remove the old default provider if available
-func (e *eventExecutor) registerDefaultProvider(provider FeatureProvider) error {
+func (e *eventExecutor) registerEventingProvider(provider FeatureProvider) {
 	e.mu.Lock()
 	defer e.mu.Unlock()
 
 	v, ok := provider.(EventHandler)
 	if !ok {
-		return nil
-	}
-
-	oldProvider := e.defaultProviderReference
-
-	// register shutdown semaphore for new default provider
-	sem := make(chan interface{})
-
-	newProvider := &providerReference{
-		eventHandler:      &v,
-		metadata:          provider.Metadata(),
-		isDefault:         true,
-		shutdownSemaphore: sem,
-	}
-
-	e.defaultProviderReference = newProvider
-	return e.listenAndShutdown(newProvider, oldProvider)
-}
-
-// registerNamedEventingProvider registers a named FeatureProvider and remove event listener for old named provider
-func (e *eventExecutor) registerNamedEventingProvider(associatedClient string, provider FeatureProvider) error {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-
-	v, ok := provider.(EventHandler)
-	if !ok {
-		return nil
-	}
-
-	oldProvider := e.namedProviderReference[associatedClient]
-
-	// register shutdown semaphore for new named provider
-	sem := make(chan interface{})
-
-	newProvider := &providerReference{
-		eventHandler:          &v,
-		clientNameAssociation: associatedClient,
-		metadata:              provider.Metadata(),
-		shutdownSemaphore:     sem,
-	}
-
-	e.namedProviderReference[associatedClient] = newProvider
-	return e.listenAndShutdown(newProvider, oldProvider)
-}
-
-// listenAndShutdown is a helper to start concurrent listening to new provider events and invoke shutdown hook of the
-// old provider
-func (e *eventExecutor) listenAndShutdown(newProvider *providerReference, oldReference *providerReference) error {
-	go func(newProvider *providerReference, oldReference *providerReference) {
-		for {
-			select {
-			case event := <-(*newProvider.eventHandler).EventChannel():
-				e.triggerEvent(event, newProvider.clientNameAssociation, newProvider.isDefault)
-			case <-newProvider.shutdownSemaphore:
-				return
-			}
-		}
-	}(newProvider, oldReference)
-
-	// shutdown old provider handling
-	if oldReference == nil {
-		return nil
-	}
-
-	select {
-	case oldReference.shutdownSemaphore <- "":
-		return nil
-	case <-time.After(200 * time.Millisecond):
-		return fmt.Errorf("event handler timeout waiting for handler shutdown")
-	}
-}
-
-// triggerEvent performs the actual event handling
-func (e *eventExecutor) triggerEvent(event Event, clientNameAssociation string, isDefault bool) {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-
-	// first run API handlers
-	for _, c := range e.apiRegistry[event.EventType] {
-		e.executeHandler(*c, event)
-	}
-
-	// then run Client handlers for name association
-
-	// first direct associates
-	associateClientRegistry := e.scopedRegistry[clientNameAssociation]
-	for _, c := range associateClientRegistry.callbacks[event.EventType] {
-		e.executeHandler(*c, event)
-	}
-
-	if !isDefault {
 		return
 	}
 
-	// handling the default provider - invoke default provider bound handlers by filtering
+	// register shutdown semaphore
+	sem := make(chan interface{})
+	e.providerShutdownHook[provider.Metadata().Name] = sem
 
-	var defaultHandlers []EventCallback
-
-	for clientName, registry := range e.scopedRegistry {
-		if _, ok := e.namedProviderReference[clientName]; !ok {
-			defaultHandlers = append(defaultHandlers, registry.callbacks[event.EventType]...)
+	go func() {
+		for {
+			select {
+			case event := <-v.EventChannel():
+				err := e.triggerEvent(event)
+				if err != nil {
+					e.logger.Error(err, fmt.Sprintf("error handling event type: %s", event.EventType))
+				}
+			case <-sem:
+				return
+			}
 		}
-	}
+	}()
+}
 
-	for _, c := range defaultHandlers {
-		e.executeHandler(*c, event)
+func (e *eventExecutor) triggerEvent(event Event) error {
+	e.mu.Lock()
+	defer e.mu.Unlock()
+
+	ctx, cancelFunc := context.WithCancel(context.Background())
+	defer cancelFunc()
+
+	group, gCtx := errgroup.WithContext(ctx)
+
+	group.Go(func() error {
+		// first run API handlers
+		for _, c := range e.apiRegistry[event.EventType] {
+			e.executeHandler(*c, "", event)
+		}
+
+		// then run Client handlers
+		// Note - we must only run associated client handlers of the provider
+		associateClientRegistry := e.scopedRegistry[event.ProviderName]
+		for _, c := range associateClientRegistry.callbacks[event.EventType] {
+			e.executeHandler(*c, associateClientRegistry.scope, event)
+		}
+
+		return nil
+	})
+
+	// wait for completion or timeout
+	select {
+	case <-time.After(handlerExecutionTime):
+		return fmt.Errorf("event handlers timeout")
+	case <-gCtx.Done():
+		return nil
 	}
 }
 
-// executeHandler is a helper which performs the actual invocation of the callback
-func (e *eventExecutor) executeHandler(f func(details EventDetails), event Event) {
-	go func() {
-		defer func() {
-			if r := recover(); r != nil {
-				e.logger.Info("recovered from a panic")
-			}
-		}()
-
-		f(EventDetails{
-			providerName: event.ProviderName,
-			ProviderEventDetails: ProviderEventDetails{
-				Message:       event.Message,
-				FlagChanges:   event.FlagChanges,
-				EventMetadata: event.EventMetadata,
-			},
-		})
+func (e *eventExecutor) executeHandler(f func(details EventDetails), clientName string, event Event) {
+	defer func() {
+		if r := recover(); r != nil {
+			e.logger.Info("recovered from a panic")
+		}
 	}()
+
+	f(EventDetails{
+		client: clientName,
+		ProviderEventDetails: ProviderEventDetails{
+			Message:       event.Message,
+			FlagChanges:   event.FlagChanges,
+			EventMetadata: event.EventMetadata,
+		},
+	})
+}
+
+func (e *eventExecutor) unregisterEventingProvider(provider FeatureProvider) error {
+	e.mu.Lock()
+	defer e.mu.Unlock()
+
+	_, ok := provider.(EventHandler)
+	if !ok {
+		return nil
+	}
+
+	sem := e.providerShutdownHook[provider.Metadata().Name]
+	if sem == nil {
+		return nil
+	}
+
+	delete(e.providerShutdownHook, provider.Metadata().Name)
+
+	// wait for completion or timeout
+	select {
+	case sem <- "":
+		return nil
+	case <-time.After(200 * time.Millisecond):
+		return fmt.Errorf("event handler of provider %s timeout with exiting", provider.Metadata().Name)
+	}
 }
